Fix a compile warning.

+++ modules/core/src/convert.dispatch.cpp
@@ -287,7 +287,7 @@ void Mat::convertTo(OutputArray dst, int type_, double alpha, double beta) const
     }
     else if( isContinuous() && dstMat.isContinuous() )
     {
-        CALL_HAL(convertScale, cv_hal_convertScale, data, 0, dstMat.data, 0, total() * cn, 1, sdepth, ddepth, alpha, beta);
+        CALL_HAL(convertScale, cv_hal_convertScale, data, 0, dstMat.data, 0, (int)total() * cn, 1, sdepth, ddepth, alpha, beta);
     }
 
     BinaryFunc func = noScale ? getConvertFunc(sdepth, ddepth) : getConvertScaleFunc(sdepth, ddepth);
